Use int32_t instead of long for stream move offset
C long is 32-bit on Windows (LLP64) and 64-bit on Linux/macOS (LP64),
so the sk_stream_move and sk_managedstream_move_proc C-API signatures
gave managed callers a silently platform-dependent parameter size. The
SkiaSharp bindings mapped long to Int32, which on Linux meant negative
offsets were misinterpreted as huge positive numbers at the ABI
boundary (the .NET marshaller zero-extends the 32-bit value into a
64-bit register slot, so -1 becomes 0x00000000FFFFFFFF when native
reads the 64-bit long).

Narrow the C-API surface to int32_t so the ABI is unambiguously 32-bit
on every platform. Internally SkManagedStream::move(long) still matches
upstream SkStream's signature; we range-check the offset before calling
into the managed MoveProc and return false on overflow rather than
silently truncating.

Practical impact: stream offsets are now explicitly capped at 2 GB on
all platforms. Windows behaviour is unchanged (it was already 32-bit);
Linux loses the theoretical ability to pass >2 GB offsets but gains
correct handling of negative offsets. The trade-off favours correctness
over a range that nobody could reliably use anyway.

+++ src/xamarin/sk_managedstream.cpp
@@ -108,7 +108,7 @@ bool dSeek(SkManagedStream* stream, void* context, size_t position) {
     if (!gProcs.fSeek) return false;
     return gProcs.fSeek(ToManagedStream(stream), context, position);
 }
-bool dMove(SkManagedStream* stream, void* context, long offset) {
+bool dMove(SkManagedStream* stream, void* context, int32_t offset) {
     if (!gProcs.fMove) return false;
     return gProcs.fMove(ToManagedStream(stream), context, offset);
 }
